Fix convert_numbers function to return integers instead of floats

+++ squares.py
@@ -39,7 +39,7 @@ def convert_numbers(list_of_strings):
     Example:
     --------
     >>> convert_numbers(["4", " 8 ", "15 16", " 23    42 "])
-    [4.0, 8.0, 15.0, 16.0, 23.0, 42.0]
+    [4, 8, 15, 16, 23, 42]
 
     """
     all_numbers = []
@@ -48,7 +48,7 @@ def convert_numbers(list_of_strings):
         # whitespace, and collect them into a single list...
         all_numbers.extend([token.strip() for token in s.split()])
     # ...then convert each substring into a number
-    return [float(number_string) for number_string in all_numbers]
+    return [int(number_string) for number_string in all_numbers]
 
 
 if __name__ == "__main__":
